Added DTR transient work around for linux and windows.

+++ src/UARTOSInterface/HardwareCOM/UOSInterface.py
@@ -20,7 +20,7 @@ class UOSInterface(metaclass=ABCMeta):
         )
 
     @abstractmethod
-    def read_response(self, timeout_s: int) -> (bool, {}):
+    def read_response(self, timeout_s: float) -> (bool, {}):
         """ Abstract method for reading ACK and Data packets from a UOSInterface.
         :param timeout_s: The maximum time this function will wait for data.
         :return: A tuple containing a success boolean at index 0 and a result-set dict at index 1.
